Move create and remove actions to main views
