+++ apps/edwoca/templates/edwoca/partials/forms/_modification_form.html
@@ -120,10 +120,9 @@
                 {% endif %}
             </div>
             <div class="flex justify-end gap-3 mt-3">
-                <form method="post" action="{% url 'edwoca:modification_delete' modification.form.instance.id %}" >
-                    {% csrf_token %}
-                    <input type="submit" name="delete_modification" value="{% translate 'remove modification' %}" class="btn btn-outline text-primary" />
-                </form>
+                <button class="btn btn-outline text-primary" value="{{ modification.form.instance.id }}" form="form" name="remove-modification">
+                    {% translate 'remove modification' %}
+                </button>
             </div>
         </div>
     </div>
