+++ apps/edwoca/urls.py
@@ -4,341 +4,246 @@ from . import views
 
 app_name='edwoca'
 
-entities = {
-        'work': '',
-        'work_title': 'work',
-        'work_bib': 'work',
-        'related_work': 'work',
-        'work_contributor': 'work',
-        'expression': 'work',
-        'expression_title': 'expression',
-        'expression_period': 'expression',
-        'expression_contributor': 'expression',
-        'manifestation': '',
-        'manifestation_contributor': 'manifestation',
-        #'manifestation_title': 'manifestation',
-        'related_manifestation': 'manifestation',
-        'item': 'manifestation',
-        'item_contributor': 'item',
-        #'item_title': 'item',
-        'provenance_state': 'item'
-    }
-
 urlpatterns = [
-        path('', views.index, name = 'index'),
-        path('works', views.WorkListView.as_view(), name = 'work_list'),
-        path('works/search', views.WorkSearchView.as_view(), name = 'work_search'),
-        path('works/new', views.work_create, name = 'work_create'),
-        path('works/<int:pk>/update', views.WorkTitleUpdateView.as_view(), name = 'work_update'),
-        path('works/<int:pk>/detail', views.WorkDetailView.as_view(), name = 'work_detail'),
-        path('works/<int:pk>/relations', views.WorkRelationsUpdateView.as_view(), name = 'work_relations'),
-        path('works/<int:pk>/relations/addrelatedwork/<int:target_related_work>', views.RelatedWorkAddView.as_view(), name = 'related_work_add'),
-        path('works/removerelatedwork/<int:pk>', views.RelatedWorkRemoveView.as_view(), name = 'related_work_remove'),
-        path('works/<int:pk>/references', views.WorkReferencesUpdateView.as_view(), name = 'work_references'),
-        path('works/<int:pk>/references/addreferencework/<int:target_reference_work>', views.reference_work_add, name = 'reference_work_add'),
-        path('works/removereferencework/<int:pk>', views.ReferenceWorkRemoveView.as_view(), name = 'reference_work_remove'),
-        path('works/<int:pk>/references/addreferenceperson/<int:target_reference_person>', views.reference_person_add, name = 'reference_person_add'),
-        path('works/removereferenceperson/<int:pk>', views.ReferencePersonRemoveView.as_view(), name = 'reference_person_remove'),
-        path('works/<int:pk>/references/addreferencecorporation/<int:target_reference_corporation>', views.reference_corporation_add, name = 'reference_corporation_add'),
-        path('works/removereferencecorporation/<int:pk>', views.ReferenceCorporationRemoveView.as_view(), name = 'reference_corporation_remove'),
-        path('works/<int:pk>/references/addreferenceplace/<int:target_reference_place>', views.reference_place_add, name = 'reference_place_add'),
-        path('works/removereferenceplace/<int:pk>', views.ReferencePlaceRemoveView.as_view(), name = 'reference_place_remove'),
-        path('works/<int:pk>/references/addreferencesubjectterm/<int:target_reference_subjectterm>', views.reference_subjectterm_add, name = 'reference_subjectterm_add'),
-        path('works/removereferencesubjectterm/<int:pk>', views.ReferenceSubjectTermRemoveView.as_view(), name = 'reference_subjectterm_remove'),
-        path('works/<int:pk>/title', views.WorkTitleUpdateView.as_view(), name = 'work_title'),
-        path('work/person_dedication/add/<int:pk>/', views.work_person_dedication_add, name='work_person_dedication_add'),
-        path('work/corporation_dedication/add/<int:pk>/', views.work_corporation_dedication_add, name='work_corporation_dedication_add'),
-        path('work/person_dedication/delete/<int:pk>/', views.work_person_dedication_delete, name='work_person_dedication_delete'),
-        path('work/corporation_dedication/delete/<int:pk>/', views.work_corporation_dedication_delete, name='work_corporation_dedication_delete'),
-        path('work/<int:pk>/person_dedication/<int:dedication_id>/add_dedicatee/<int:person_id>/', views.work_person_dedication_add_dedicatee, name='work_person_dedication_add_dedicatee'),
-        path('work/<int:pk>/person_dedication/<int:dedication_id>/remove_dedicatee/', views.work_person_dedication_remove_dedicatee, name='work_person_dedication_remove_dedicatee'),
-        path('work/<int:pk>/corporation_dedication/<int:dedication_id>/add_dedicatee/<int:corporation_id>/', views.work_corporation_dedication_add_dedicatee, name='work_corporation_dedication_add_dedicatee'),
-        path('work/<int:pk>/corporation_dedication/<int:dedication_id>/remove_dedicatee/', views.work_corporation_dedication_remove_dedicatee, name='work_corporation_dedication_remove_dedicatee'),
-        path('work/<int:pk>/dedication/<int:dedication_id>/add_place/<int:place_id>/', views.work_dedication_add_place, name='work_dedication_add_place'),
-        path('work/<int:pk>/dedication/<int:dedication_id>/remove_place/', views.work_dedication_remove_place, name='work_dedication_remove_place'),
+    path('', views.index, name = 'index'),
+    path('works', views.WorkListView.as_view(), name = 'work_list'),
+    path('works/search', views.WorkSearchView.as_view(), name = 'work_search'),
+    path('works/new', views.work_create, name = 'work_create'),
+    path('works/<int:pk>/update', views.WorkTitleUpdateView.as_view(), name = 'work_update'),
+    path('works/<int:pk>/detail', views.WorkDetailView.as_view(), name = 'work_detail'),
+    path('works/<int:pk>/relations', views.WorkRelationsUpdateView.as_view(), name = 'work_relations'),
+    path('works/<int:pk>/relations/addrelatedwork/<int:target_related_work>', views.RelatedWorkAddView.as_view(), name = 'related_work_add'),
+    path('works/removerelatedwork/<int:pk>', views.RelatedWorkRemoveView.as_view(), name = 'related_work_remove'),
+    path('works/<int:pk>/references', views.WorkReferencesUpdateView.as_view(), name = 'work_references'),
+    path('works/<int:pk>/references/addreferencework/<int:target_reference_work>', views.reference_work_add, name = 'reference_work_add'),
+    path('works/removereferencework/<int:pk>', views.ReferenceWorkRemoveView.as_view(), name = 'reference_work_remove'),
+    path('works/<int:pk>/references/addreferenceperson/<int:target_reference_person>', views.reference_person_add, name = 'reference_person_add'),
+    path('works/removereferenceperson/<int:pk>', views.ReferencePersonRemoveView.as_view(), name = 'reference_person_remove'),
+    path('works/<int:pk>/references/addreferencecorporation/<int:target_reference_corporation>', views.reference_corporation_add, name = 'reference_corporation_add'),
+    path('works/removereferencecorporation/<int:pk>', views.ReferenceCorporationRemoveView.as_view(), name = 'reference_corporation_remove'),
+    path('works/<int:pk>/references/addreferenceplace/<int:target_reference_place>', views.reference_place_add, name = 'reference_place_add'),
+    path('works/removereferenceplace/<int:pk>', views.ReferencePlaceRemoveView.as_view(), name = 'reference_place_remove'),
+    path('works/<int:pk>/references/addreferencesubjectterm/<int:target_reference_subjectterm>', views.reference_subjectterm_add, name = 'reference_subjectterm_add'),
+    path('works/removereferencesubjectterm/<int:pk>', views.ReferenceSubjectTermRemoveView.as_view(), name = 'reference_subjectterm_remove'),
+    path('works/<int:pk>/title', views.WorkTitleUpdateView.as_view(), name = 'work_title'),
+    path('work/person_dedication/add/<int:pk>/', views.work_person_dedication_add, name='work_person_dedication_add'),
+    path('work/corporation_dedication/add/<int:pk>/', views.work_corporation_dedication_add, name='work_corporation_dedication_add'),
+    path('work/person_dedication/delete/<int:pk>/', views.work_person_dedication_delete, name='work_person_dedication_delete'),
+    path('work/corporation_dedication/delete/<int:pk>/', views.work_corporation_dedication_delete, name='work_corporation_dedication_delete'),
+    path('work/<int:pk>/person_dedication/<int:dedication_id>/add_dedicatee/<int:person_id>/', views.work_person_dedication_add_dedicatee, name='work_person_dedication_add_dedicatee'),
+    path('work/<int:pk>/person_dedication/<int:dedication_id>/remove_dedicatee/', views.work_person_dedication_remove_dedicatee, name='work_person_dedication_remove_dedicatee'),
+    path('work/<int:pk>/corporation_dedication/<int:dedication_id>/add_dedicatee/<int:corporation_id>/', views.work_corporation_dedication_add_dedicatee, name='work_corporation_dedication_add_dedicatee'),
+    path('work/<int:pk>/corporation_dedication/<int:dedication_id>/remove_dedicatee/', views.work_corporation_dedication_remove_dedicatee, name='work_corporation_dedication_remove_dedicatee'),
+    path('work/<int:pk>/dedication/<int:dedication_id>/add_place/<int:place_id>/', views.work_dedication_add_place, name='work_dedication_add_place'),
+    path('work/<int:pk>/dedication/<int:dedication_id>/remove_place/', views.work_dedication_remove_place, name='work_dedication_remove_place'),
 
-        path('works/<int:pk>/contributors', views.WorkContributorsUpdateView.as_view(), name = 'work_contributors'),
-        path('works/<int:pk>/contributors/addcontributor/<int:person>', views.WorkContributorAddView.as_view(), name = 'work_contributor_add'),
-        path('works/removecontributor/<int:pk>', views.WorkContributorRemoveView.as_view(), name = 'work_contributor_remove'),
-        path('works/<int:pk>/history', views.WorkHistoryUpdateView.as_view(), name = 'work_history'),
-        path('works/<int:pk>/bibliography', views.WorkBibliographyUpdateView.as_view(), name = 'work_bibliography'),
-        path('works/<int:pk>/bibliography/addbib/<str:zotitem_key>', views.WorkBibAddView.as_view(), name = 'work_bib_add'),
-        path('works/removebib/<int:pk>', views.WorkBibDeleteView.as_view(), name = 'work_bib_remove'),
-        path('works/<int:pk>/letteradd/<int:letter_pk>', views.work_letter_add, name = 'work_letter_add'),
-        path('works/<int:pk>/letterremove/<int:letter_pk>', views.work_letter_remove, name = 'work_letter_remove'),
-        path('works/<int:pk>/comment', views.WorkCommentUpdateView.as_view(), name = 'work_comment'),
-        path('works/<int:pk>/head_comment', views.WorkHeadCommentUpdateView.as_view(), name = 'work_head_comment'),
-        path('works/<int:pk>/delete', views.WorkDeleteView.as_view(), name = 'work_delete'),
-        path('expressions', views.ExpressionListView.as_view(), name = 'expression_list'),
-        path('expressions/search', views.ExpressionSearchView.as_view(), name = 'expression_search'),
-        path('expressions/<int:pk>/update', views.expression_update, name = 'expression_update'),
-        path('expressions/<int:pk>/detail', views.ExpressionDetailView.as_view(), name = 'expression_detail'),
-        path('expressions/<int:pk>/index_number_create', views.index_number_create, name = 'index_number_create'),
-        path('index_number/<int:pk>/delete', views.IndexNumberDeleteView.as_view(), name = 'index_number_delete'),
-        path('expressions/<int:pk>/delete', views.ExpressionDeleteView.as_view(), name = 'expression_delete'),
-        path('expressions/<int:pk>/title', views.ExpressionTitleUpdateView.as_view(), name = 'expression_title'),
-        path('expressions/<int:pk>/relations', views.ExpressionRelationsUpdateView.as_view(), name = 'expression_relations'),
-        path('expressions/<int:pk>/relations/addrelatedexpression/<int:target_expression>', views.RelatedExpressionAddView.as_view(), name = 'related_expression_add'),
-        path('expressions/removerelatedexpression/<int:pk>', views.RelatedExpressionRemoveView.as_view(), name = 'related_expression_remove'),
-        path('expressions/<int:pk>/contributors', views.ExpressionContributorsUpdateView.as_view(), name = 'expression_contributors'),
-        path('expressions/<int:pk>/contributors/addcontributor/<int:person>', views.ExpressionContributorAddView.as_view(), name = 'expression_contributor_add'),
-        path('expressions/removecontributor/<int:pk>', views.ExpressionContributorRemoveView.as_view(), name = 'expression_contributor_remove'),
-        path('expressions/<int:pk>/history', views.ExpressionHistoryUpdateView.as_view(), name = 'expression_history'),
-        path('expressions/<int:pk>/categorisation', views.ExpressionCategorisationUpdateView.as_view(), name = 'expression_categorisation'),
-        path('expressions/<int:pk>/mediumofperformance', views.ExpressionMediumofperformanceUpdateView.as_view(), name = 'expression_mediumofperformance'),
-        path('expressions/<int:pk>/movements', views.ExpressionMovementsUpdateView.as_view(), name = 'expression_movements'),
-        path('expressions/<int:pk>/comment', views.ExpressionCommentUpdateView.as_view(), name = 'expression_comment'),
-        path('expressions/<int:pk>/<str:direction>/swap', views.expression_swap_view, name='expression_swap_order'),
-        #path('expressions/<int:work_id>/createexpression', views.ExpressionCreateView.as_view(), name = 'expression_create'),
-        path('work/<int:pk>/createexpression', views.work_expression_create, name = 'expression_create'),
-        path('expressions/<int:pk>/delete', views.ExpressionDeleteView.as_view(), name = 'expression_delete'),
-        path('composites', views.CompositeListView.as_view(), name = 'composite_list'),
-        path('composites/search', views.CompositeSearchView.as_view(), name = 'composite_search'),
-        path('composites/<int:pk>/delete', views.CompositeDeleteView.as_view(), name = 'composite_delete'),
-        path('composites/new', views.composite_create, name = 'composite_create'),
-        path('composites/<int:pk>/title', views.composite_title_update, name = 'composite_title'),
-        path('composite_title/<int:pk>/delete', views.CompositeTitleDeleteView.as_view(), name = 'composite_title_delete'),
-        path('composites/<int:pk>/add_title', views.composite_title_create, name = 'composite_title_add'),
-        path('composites/<int:pk>/relations', views.CompositeRelationsUpdateView.as_view(), name = 'composite_relations'),
-        path('composites/<int:pk>/history/', views.CompositeHistoryUpdateView.as_view(), name='composite_history'),
-        path('composites/<int:pk>/provenance/', views.composite_provenance, name='composite_provenance'),
-        path('singletons', views.SingletonListView.as_view(), name = 'singleton_list'),
-        path('singletons/search', views.SingletonSearchView.as_view(), name = 'singleton_search'),
-        path('singletons/new', views.singleton_create, name = 'singleton_create'),
-        path('manifestations/<int:pk>/create_part', views.part_create, name = 'part_create'),
-        path('manifestations/<int:pk>/create_part/publisher/<int:publisher_pk>', views.part_create, name = 'part_create'),
-        path('manifestations/<int:pk>/create_component', views.component_create, name = 'component_create'),
-        path('manifestations/<int:pk>/create_component/publisher/<int:publisher_pk>', views.component_create, name = 'component_create'),
-        path('manifestations/<int:pk>/create_singleton_part/<str:part_label>', views.singleton_part_create, name = 'singleton_part_create'),
-        path('manifestations/<int:pk>/create_singleton_component', views.singleton_component_create, name = 'singleton_component_create'),
-        path('singletoncollections/new', views.singleton_collection_create, name = 'singleton_collection_create'),
-        path('manifestations', views.ManifestationListView.as_view(), name = 'manifestation_list'),
-        path('manifestations/search', views.ManifestationSearchView.as_view(), name = 'manifestation_search'),
-        path('manifestations/<int:pk>/update', views.manifestation_update, name = 'manifestation_update'),
-        path('manifestations/<int:pk>/detail', views.ManifestationDetailView.as_view(), name = 'manifestation_detail'),
-        path('manifestations/<int:pk>/set_singleton', views.manifestation_set_singleton, name = 'manifestation_set_singleton'),
-        path('manifestations/<int:pk>/unset_singleton', views.manifestation_unset_singleton, name = 'manifestation_unset_singleton'),
-        path('manifestations/<int:pk>/set_collection', views.manifestation_set_collection, name = 'manifestation_set_collection'),
-        path('manifestations/<int:pk>/unset_collection', views.manifestation_unset_collection, name = 'manifestation_unset_collection'),
-        path('manifestations/<int:pk>/set_missing', views.manifestation_set_missing, name = 'manifestation_set_missing'),
-        path('manifestations/<int:pk>/unset_missing', views.manifestation_unset_missing, name = 'manifestation_unset_missing'),
-        path('manifestations/new', views.manifestation_create, name = 'manifestation_create'),
-        path('manifestationcollections/new', views.manifestation_collection_create, name = 'manifestation_collection_create'),
-        path('manifestations/new/publisher/<int:publisher_pk>', views.manifestation_create, name = 'manifestation_create'),
-        path('manifestationcollections/new/publisher/<int:publisher_pk>', views.manifestation_collection_create, name = 'manifestation_collection_create'),
-        path('manifestations/<int:pk>/delete', views.ManifestationDeleteView.as_view(), name = 'manifestation_delete'),
-        path('manifestations/<int:pk>/copy', views.manifestation_copy, name = 'manifestation_copy'),
-        path('manifestations/<int:pk>/title', views.manifestation_title_update, name = 'manifestation_title'),
-        path('manifestation_title/<int:pk>/delete', views.ManifestationTitleDeleteView.as_view(), name = 'manifestation_title_delete'),
-        path('manifestation_title_handwriting/<int:pk>/delete', views.ManifestationTitleHandwritingDeleteView.as_view(), name = 'manifestation_title_handwriting_delete'),
-        path('manifestations/<int:pk>/title/<int:title_id>/addwriter/<int:writer_id>', views.manifestation_writer_add, name = 'manifestation_writer_add'),
-        path('manifestations/<int:pk>/title_handwriting/<int:title_handwriting_pk>/add_writer/<int:person_pk>', views.manifestation_title_add_handwriting_writer, name = 'manifestation_title_add_handwriting_writer'),
-        path('manifestations/<int:pk>/title_handwriting/<int:title_handwriting_pk>/remove_writer', views.manifestation_title_remove_handwriting_writer, name = 'manifestation_title_remove_handwriting_writer'),
-        path('manifestations/<int:pk>/title/<int:title_id>/removewriter', views.manifestation_writer_remove, name = 'manifestation_writer_remove'),
-        path('manifestations/<int:pk>/add_title', views.manifestation_title_create, name = 'manifestation_title_add'),
-        path('manifestations/<int:pk>/add_dedicatee/<int:person_id>', views.manifestation_add_dedicatee, name = 'manifestation_add_dedicatee'),
-        path('manifestations/<int:pk>/remove_dedicatee/<int:dedicatee_id>', views.manifestation_remove_dedicatee, name = 'manifestation_remove_dedicatee'),
-        path('manifestations/<int:pk>/relations', views.ManifestationRelationsUpdateView.as_view(), name = 'manifestation_relations'),
-        path('manifestations/<int:pk>/add_expression/<int:expression_pk>', views.manifestation_expression_add, name = 'manifestation_expression_add'),
-        path('manifestations/<int:pk>/remove_expression/<int:expression_pk>', views.manifestation_expression_remove, name = 'manifestation_expression_remove'),
-        path('manifestations/<int:pk>/relations/addrelatedmanifestation/<int:target_related_manifestation>', views.RelatedManifestationAddView.as_view(), name = 'related_manifestation_add'),
-        path('manifestations/<int:referrer_pk>/removerelatedmanifestation/<int:pk>', views.RelatedManifestationRemoveView.as_view(), name = 'related_manifestation_remove'),
-        path('manifestations/removecollection/<int:pk>', views.collection_remove, name = 'collection_remove'),
-        path('manifestations/<int:pk>/history/', views.ManifestationHistoryUpdateView.as_view(), name='manifestation_history'),
-        path('manifestations/<int:pk>/provenance/', views.ManifestationProvenanceView.as_view(), name='manifestation_provenance'),
-        path('manifestations/<int:pk>/item_create', views.manifestation_item_create, name = 'manifestation_item_create'),
-        path('manifestations/autocomplete/', views.manifestation_autocomplete, name = 'manifestation_autocomplete'),
-        path('manifestation/item/<int:pk>/move-modal', views.item_move_modal, name='item_move_modal'),
-        path('manifestation/item/<int:item_pk>/move-execute/', views.item_move_view, name = 'item_move_execute'),
-        path('manifestations/item_delete/<int:item_id>', views.manifestation_item_delete, name = 'manifestation_item_delete'),
-        path('item/<int:item_id>/provenance/person/add/', views.person_provenance_add, name='person_provenance_add'),
-        path('item/provenance/person/delete/<int:pk>/', views.PersonProvenanceStationDeleteView.as_view(), name='person_provenance_delete'),
-        path('item/<int:item_id>/provenance/corporation/add/', views.corporation_provenance_add, name='corporation_provenance_add'),
-        path('item/provenance/corporation/delete/<int:pk>/', views.CorporationProvenanceStationDeleteView.as_view(), name='corporation_provenance_delete'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/add_owner/<int:person_id>/', views.person_provenance_add_owner, name='person_provenance_add_owner'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/add_bib/<str:bib_id>/', views.person_provenance_add_bib, name='person_provenance_add_bib'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_webref/', views.corporation_provenance_add_webref, name='corporation_provenance_add_webref'),
-        path('item/provenance/corporation/remove_webref/<str:webref_id>', views.corporation_provenance_remove_webref, name='corporation_provenance_remove_webref'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/add_webref/', views.person_provenance_add_webref, name='person_provenance_add_webref'),
-        path('item/provenance/person/remove_webref/<str:webref_id>', views.person_provenance_remove_webref, name='person_provenance_remove_webref'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_owner/<int:corporation_id>/', views.corporation_provenance_add_owner, name='corporation_provenance_add_owner'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_bib/<str:bib_id>/', views.corporation_provenance_add_bib, name='corporation_provenance_add_bib'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/remove_owner/<int:owner_id>', views.person_provenance_remove_owner, name='person_provenance_remove_owner'),
-        path('item/provenance/person/remove_bib/<str:bib_id>', views.person_provenance_remove_bib, name='person_provenance_remove_bib'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/add_letter/<int:letter_pk>/', views.person_provenance_add_letter, name='person_provenance_add_letter'),
-        path('item/<int:pk>/provenance/person/<int:pps_id>/remove_letter/<int:letter_pk>/', views.person_provenance_remove_letter, name='person_provenance_remove_letter'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/remove_owner/', views.corporation_provenance_remove_owner, name='corporation_provenance_remove_owner'),
-        path('item/provenance/corporation/remove_bib/<str:bib_id>', views.corporation_provenance_remove_bib, name='corporation_provenance_remove_bib'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_letter/<int:letter_pk>/', views.corporation_provenance_add_letter, name='corporation_provenance_add_letter'),
-        path('item/<int:pk>/<str:direction>/swap', views.item_swap_view, name='item_swap_order'),
-        path('item/<int:pk>/provenance/corporation/<int:cps_id>/remove_letter/<int:letter_pk>/', views.corporation_provenance_remove_letter, name='corporation_provenance_remove_letter'),
-        path('item/remove_signature/<int:pk>', views.ItemSignatureDeleteView.as_view(), name='item_signature_remove'),
-        path('manifestations/<int:pk>/bibliography', views.ManifestationBibliographyUpdateView.as_view(), name = 'manifestation_bibliography'),
-        path('manifestations/<int:pk>/bibliography/addbib/<str:zotitem_key>', views.ManifestationBibAddView.as_view(), name = 'manifestation_bib_add'),
-        path('manifestations/removebib/<int:pk>', views.ManifestationBibDeleteView.as_view(), name = 'manifestation_bib_remove'),
-        path('manifestations/<int:pk>/letteradd/<int:letter_pk>', views.manifestation_letter_add, name = 'manifestation_letter_add'),
-        path('manifestations/<int:pk>/letterremove/<int:letter_pk>', views.manifestation_letter_remove, name = 'manifestation_letter_remove'),
-        path('manifestations/<int:pk>/comment', views.ManifestationCommentUpdateView.as_view(), name = 'manifestation_comment'),
-        path('manifestations/<int:pk>/classification', views.ManifestationClassificationUpdateView.as_view(), name = 'manifestation_classification'),
-        path('manifestations/<int:pk>/print', views.manifestation_print_update, name = 'manifestation_print'),
-        path('manifestations/<int:pk>/print/add_publication_publisher/<int:publication_id>/<int:publisher_id>/', views.manifestation_add_publication_publisher, name = 'manifestation_add_publication_publisher'),
-        path('manifestations/<int:pk>/print/remove_publication_publisher/<int:publication_id>/', views.manifestation_remove_publication_publisher, name = 'manifestation_remove_publication_publisher'),
-        path('manifestations/<int:pk>/print/add_publication_place/<int:publication_id>/<int:place_id>/', views.manifestation_add_publication_place, name = 'manifestation_add_publication_place'),
-        path('manifestations/<int:pk>/print/remove_publication_place/<int:publication_id>/<int:place_id>', views.manifestation_remove_publication_place, name = 'manifestation_remove_publication_place'),
-        path('manifestations/publication/<int:pk>/delete/', views.ManifestationPublicationDeleteView.as_view(), name = 'manifestation_publication_delete'),
-        path('manifestations/<int:pk>/manuscript', views.manifestation_manuscript_update, name = 'manifestation_manuscript'),
-        path('manifestations/modification/<int:pk>/delete', views.ModificationDeleteView.as_view(), name='modification_delete'),
-        #path('manifestations/modification_handwriting/<int:pk>/delete', views.ManifestationModificationHandwritingDeleteView.as_view(), name='manifestation_modification_handwriting_delete'),
-        #path('manifestations/modification_handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.modification_add_handwriting_writer, name='modification_add_handwriting_writer'),
-        #path('manifestations/modification_handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.modification_remove_handwriting_writer, name='modification_remove_handwriting_writer'),
-        #path('manifestations/modification_handwriting/<int:handwriting_pk>/remove_writer', views.modification_remove_handwriting_writer, name='modification_remove_handwriting_writer'),
-        path('manifestations/modification/<int:modification_pk>/add_related_work/<int:work_pk>', views.modification_add_related_work, name='manifestation_modification_add_related_work'),
-        path('manifestations/modification/<int:modification_pk>/remove_related_work', views.modification_remove_related_work, name='manifestation_modification_remove_related_work'),
-        path('manifestations/modification/<int:modification_pk>/add_related_expression/<int:expression_pk>', views.modification_add_related_expression, name='manifestation_modification_add_related_expression'),
-        path('manifestations/modification/<int:modification_pk>/remove_related_expression', views.modification_remove_related_expression, name='manifestation_modification_remove_related_expression'),
-        path('manifestations/<int:pk>/digital_copy/', views.manifestation_digital_copy, name='manifestation_digital_copy'),
-        path('manifestations/<int:pk>/digital_copy/add/', views.manifestation_digital_copy_add, name='manifestation_digital_copy_add'),
-        path('manifestations/digital_copy/delete/<int:pk>/', views.ManifestationDigitalCopyDeleteView.as_view(), name='manifestation_digital_copy_delete'),
-        path('manifestations/<int:pk>/handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.manifestation_add_handwriting_writer, name = 'manifestation_add_handwriting_writer'),
-        path('manifestations/<int:pk>/handwriting/<int:handwriting_pk>/remove_writer', views.manifestation_remove_handwriting_writer, name = 'manifestation_remove_handwriting_writer'),
-        path('manifestations/handwriting/<int:pk>/delete', views.ManifestationHandwritingDeleteView.as_view(), name = 'manifestation_handwriting_delete'),
-        path('manifestations/modificationhandwriting/<int:pk>/delete', views.ModificationHandwritingDeleteView.as_view(), name = 'modification_handwriting_delete'),
-        path('manifestations/<int:pk>/add_publisher/<int:publisher_id>', views.manifestation_add_publisher, name = 'manifestation_add_publisher'),
-        path('manifestations/<int:pk>/remove_publisher', views.manifestation_remove_publisher, name = 'manifestation_remove_publisher'),
-        path('manifestations/<int:pk>/add_stitcher/<int:stitcher_id>', views.manifestation_add_stitcher, name = 'manifestation_add_stitcher'),
-        path('manifestations/<int:pk>/remove_stitcher', views.manifestation_remove_stitcher, name = 'manifestation_remove_stitcher'),
+    path('works/<int:pk>/contributors', views.WorkContributorsUpdateView.as_view(), name = 'work_contributors'),
+    path('works/<int:pk>/contributors/addcontributor/<int:person>', views.WorkContributorAddView.as_view(), name = 'work_contributor_add'),
+    path('works/removecontributor/<int:pk>', views.WorkContributorRemoveView.as_view(), name = 'work_contributor_remove'),
+    path('works/<int:pk>/history', views.WorkHistoryUpdateView.as_view(), name = 'work_history'),
+    path('works/<int:pk>/bibliography', views.WorkBibliographyUpdateView.as_view(), name = 'work_bibliography'),
+    path('works/<int:pk>/bibliography/addbib/<str:zotitem_key>', views.WorkBibAddView.as_view(), name = 'work_bib_add'),
+    path('works/removebib/<int:pk>', views.WorkBibDeleteView.as_view(), name = 'work_bib_remove'),
+    path('works/<int:pk>/letteradd/<int:letter_pk>', views.work_letter_add, name = 'work_letter_add'),
+    path('works/<int:pk>/letterremove/<int:letter_pk>', views.work_letter_remove, name = 'work_letter_remove'),
+    path('works/<int:pk>/comment', views.WorkCommentUpdateView.as_view(), name = 'work_comment'),
+    path('works/<int:pk>/head_comment', views.WorkHeadCommentUpdateView.as_view(), name = 'work_head_comment'),
+    path('works/<int:pk>/delete', views.WorkDeleteView.as_view(), name = 'work_delete'),
+    path('expressions', views.ExpressionListView.as_view(), name = 'expression_list'),
+    path('expressions/search', views.ExpressionSearchView.as_view(), name = 'expression_search'),
+    path('expressions/<int:pk>/update', views.expression_update, name = 'expression_update'),
+    path('expressions/<int:pk>/detail', views.ExpressionDetailView.as_view(), name = 'expression_detail'),
+    path('expressions/<int:pk>/index_number_create', views.index_number_create, name = 'index_number_create'),
+    path('index_number/<int:pk>/delete', views.IndexNumberDeleteView.as_view(), name = 'index_number_delete'),
+    path('expressions/<int:pk>/delete', views.ExpressionDeleteView.as_view(), name = 'expression_delete'),
+    path('expressions/<int:pk>/title', views.ExpressionTitleUpdateView.as_view(), name = 'expression_title'),
+    path('expressions/<int:pk>/relations', views.ExpressionRelationsUpdateView.as_view(), name = 'expression_relations'),
+    path('expressions/<int:pk>/relations/addrelatedexpression/<int:target_expression>', views.RelatedExpressionAddView.as_view(), name = 'related_expression_add'),
+    path('expressions/removerelatedexpression/<int:pk>', views.RelatedExpressionRemoveView.as_view(), name = 'related_expression_remove'),
+    path('expressions/<int:pk>/contributors', views.ExpressionContributorsUpdateView.as_view(), name = 'expression_contributors'),
+    path('expressions/<int:pk>/contributors/addcontributor/<int:person>', views.ExpressionContributorAddView.as_view(), name = 'expression_contributor_add'),
+    path('expressions/removecontributor/<int:pk>', views.ExpressionContributorRemoveView.as_view(), name = 'expression_contributor_remove'),
+    path('expressions/<int:pk>/history', views.ExpressionHistoryUpdateView.as_view(), name = 'expression_history'),
+    path('expressions/<int:pk>/categorisation', views.ExpressionCategorisationUpdateView.as_view(), name = 'expression_categorisation'),
+    path('expressions/<int:pk>/mediumofperformance', views.ExpressionMediumofperformanceUpdateView.as_view(), name = 'expression_mediumofperformance'),
+    path('expressions/<int:pk>/movements', views.ExpressionMovementsUpdateView.as_view(), name = 'expression_movements'),
+    path('expressions/<int:pk>/comment', views.ExpressionCommentUpdateView.as_view(), name = 'expression_comment'),
+    path('expressions/<int:pk>/<str:direction>/swap', views.expression_swap_view, name='expression_swap_order'),
+    path('work/<int:pk>/createexpression', views.work_expression_create, name = 'expression_create'),
+    path('expressions/<int:pk>/delete', views.ExpressionDeleteView.as_view(), name = 'expression_delete'),
+    path('composites', views.CompositeListView.as_view(), name = 'composite_list'),
+    path('composites/search', views.CompositeSearchView.as_view(), name = 'composite_search'),
+    path('composites/<int:pk>/delete', views.CompositeDeleteView.as_view(), name = 'composite_delete'),
+    path('composites/new', views.composite_create, name = 'composite_create'),
+    path('composites/<int:pk>/title', views.composite_title_update, name = 'composite_title'),
+    path('composite_title/<int:pk>/delete', views.CompositeTitleDeleteView.as_view(), name = 'composite_title_delete'),
+    path('composites/<int:pk>/add_title', views.composite_title_create, name = 'composite_title_add'),
+    path('composites/<int:pk>/relations', views.CompositeRelationsUpdateView.as_view(), name = 'composite_relations'),
+    path('composites/<int:pk>/history/', views.CompositeHistoryUpdateView.as_view(), name='composite_history'),
+    path('composites/<int:pk>/provenance/', views.composite_provenance, name='composite_provenance'),
+    path('singletons', views.SingletonListView.as_view(), name = 'singleton_list'),
+    path('singletons/search', views.SingletonSearchView.as_view(), name = 'singleton_search'),
+    path('singletons/new', views.singleton_create, name = 'singleton_create'),
+    path('manifestations/<int:pk>/create_part', views.part_create, name = 'part_create'),
+    path('manifestations/<int:pk>/create_part/publisher/<int:publisher_pk>', views.part_create, name = 'part_create'),
+    path('manifestations/<int:pk>/create_component', views.component_create, name = 'component_create'),
+    path('manifestations/<int:pk>/create_component/publisher/<int:publisher_pk>', views.component_create, name = 'component_create'),
+    path('manifestations/<int:pk>/create_singleton_part/<str:part_label>', views.singleton_part_create, name = 'singleton_part_create'),
+    path('manifestations/<int:pk>/create_singleton_component', views.singleton_component_create, name = 'singleton_component_create'),
+    path('singletoncollections/new', views.singleton_collection_create, name = 'singleton_collection_create'),
+    path('manifestations', views.ManifestationListView.as_view(), name = 'manifestation_list'),
+    path('manifestations/search', views.ManifestationSearchView.as_view(), name = 'manifestation_search'),
+    path('manifestations/<int:pk>/update', views.manifestation_update, name = 'manifestation_update'),
+    path('manifestations/<int:pk>/detail', views.ManifestationDetailView.as_view(), name = 'manifestation_detail'),
+    path('manifestations/<int:pk>/set_singleton', views.manifestation_set_singleton, name = 'manifestation_set_singleton'),
+    path('manifestations/<int:pk>/unset_singleton', views.manifestation_unset_singleton, name = 'manifestation_unset_singleton'),
+    path('manifestations/<int:pk>/set_collection', views.manifestation_set_collection, name = 'manifestation_set_collection'),
+    path('manifestations/<int:pk>/unset_collection', views.manifestation_unset_collection, name = 'manifestation_unset_collection'),
+    path('manifestations/<int:pk>/set_missing', views.manifestation_set_missing, name = 'manifestation_set_missing'),
+    path('manifestations/<int:pk>/unset_missing', views.manifestation_unset_missing, name = 'manifestation_unset_missing'),
+    path('manifestations/new', views.manifestation_create, name = 'manifestation_create'),
+    path('manifestationcollections/new', views.manifestation_collection_create, name = 'manifestation_collection_create'),
+    path('manifestations/new/publisher/<int:publisher_pk>', views.manifestation_create, name = 'manifestation_create'),
+    path('manifestationcollections/new/publisher/<int:publisher_pk>', views.manifestation_collection_create, name = 'manifestation_collection_create'),
+    path('manifestations/<int:pk>/delete', views.ManifestationDeleteView.as_view(), name = 'manifestation_delete'),
+    path('manifestations/<int:pk>/copy', views.manifestation_copy, name = 'manifestation_copy'),
+    path('manifestations/<int:pk>/title', views.manifestation_title_update, name = 'manifestation_title'),
+    path('manifestations/<int:pk>/title_handwriting/<int:title_handwriting_pk>/add_writer/<int:person_pk>', views.manifestation_title_add_handwriting_writer, name = 'manifestation_title_add_handwriting_writer'),
+    path('manifestations/<int:pk>/title_handwriting/<int:title_handwriting_pk>/remove_writer', views.manifestation_title_remove_handwriting_writer, name = 'manifestation_title_remove_handwriting_writer'),
+    path('manifestations/<int:pk>/relations', views.ManifestationRelationsUpdateView.as_view(), name = 'manifestation_relations'),
+    path('manifestations/<int:pk>/add_expression/<int:expression_pk>', views.manifestation_expression_add, name = 'manifestation_expression_add'),
+    path('manifestations/<int:pk>/remove_expression/<int:expression_pk>', views.manifestation_expression_remove, name = 'manifestation_expression_remove'),
+    path('manifestations/<int:pk>/relations/addrelatedmanifestation/<int:target_related_manifestation>', views.RelatedManifestationAddView.as_view(), name = 'related_manifestation_add'),
+    path('manifestations/<int:referrer_pk>/removerelatedmanifestation/<int:pk>', views.RelatedManifestationRemoveView.as_view(), name = 'related_manifestation_remove'),
+    path('manifestations/removecollection/<int:pk>', views.collection_remove, name = 'collection_remove'),
+    path('manifestations/<int:pk>/history/', views.ManifestationHistoryUpdateView.as_view(), name='manifestation_history'),
+    path('manifestations/<int:pk>/provenance/', views.ManifestationProvenanceView.as_view(), name='manifestation_provenance'),
+    path('manifestations/<int:pk>/item_create', views.manifestation_item_create, name = 'manifestation_item_create'),
+    path('manifestations/autocomplete/', views.manifestation_autocomplete, name = 'manifestation_autocomplete'),
+    path('manifestation/item/<int:pk>/move-modal', views.item_move_modal, name='item_move_modal'),
+    path('manifestation/item/<int:item_pk>/move-execute/', views.item_move_view, name = 'item_move_execute'),
+    path('manifestations/item_delete/<int:item_id>', views.manifestation_item_delete, name = 'manifestation_item_delete'),
+    path('item/<int:item_id>/provenance/person/add/', views.person_provenance_add, name='person_provenance_add'),
+    path('item/provenance/person/delete/<int:pk>/', views.PersonProvenanceStationDeleteView.as_view(), name='person_provenance_delete'),
+    path('item/<int:item_id>/provenance/corporation/add/', views.corporation_provenance_add, name='corporation_provenance_add'),
+    path('item/provenance/corporation/delete/<int:pk>/', views.CorporationProvenanceStationDeleteView.as_view(), name='corporation_provenance_delete'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/add_owner/<int:person_id>/', views.person_provenance_add_owner, name='person_provenance_add_owner'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/add_bib/<str:bib_id>/', views.person_provenance_add_bib, name='person_provenance_add_bib'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_webref/', views.corporation_provenance_add_webref, name='corporation_provenance_add_webref'),
+    path('item/provenance/corporation/remove_webref/<str:webref_id>', views.corporation_provenance_remove_webref, name='corporation_provenance_remove_webref'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/add_webref/', views.person_provenance_add_webref, name='person_provenance_add_webref'),
+    path('item/provenance/person/remove_webref/<str:webref_id>', views.person_provenance_remove_webref, name='person_provenance_remove_webref'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_owner/<int:corporation_id>/', views.corporation_provenance_add_owner, name='corporation_provenance_add_owner'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_bib/<str:bib_id>/', views.corporation_provenance_add_bib, name='corporation_provenance_add_bib'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/remove_owner/<int:owner_id>', views.person_provenance_remove_owner, name='person_provenance_remove_owner'),
+    path('item/provenance/person/remove_bib/<str:bib_id>', views.person_provenance_remove_bib, name='person_provenance_remove_bib'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/add_letter/<int:letter_pk>/', views.person_provenance_add_letter, name='person_provenance_add_letter'),
+    path('item/<int:pk>/provenance/person/<int:pps_id>/remove_letter/<int:letter_pk>/', views.person_provenance_remove_letter, name='person_provenance_remove_letter'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/remove_owner/', views.corporation_provenance_remove_owner, name='corporation_provenance_remove_owner'),
+    path('item/provenance/corporation/remove_bib/<str:bib_id>', views.corporation_provenance_remove_bib, name='corporation_provenance_remove_bib'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/add_letter/<int:letter_pk>/', views.corporation_provenance_add_letter, name='corporation_provenance_add_letter'),
+    path('item/<int:pk>/<str:direction>/swap', views.item_swap_view, name='item_swap_order'),
+    path('item/<int:pk>/provenance/corporation/<int:cps_id>/remove_letter/<int:letter_pk>/', views.corporation_provenance_remove_letter, name='corporation_provenance_remove_letter'),
+    path('manifestations/<int:pk>/bibliography', views.ManifestationBibliographyUpdateView.as_view(), name = 'manifestation_bibliography'),
+    path('manifestations/<int:pk>/bibliography/addbib/<str:zotitem_key>', views.ManifestationBibAddView.as_view(), name = 'manifestation_bib_add'),
+    path('manifestations/removebib/<int:pk>', views.ManifestationBibDeleteView.as_view(), name = 'manifestation_bib_remove'),
+    path('manifestations/<int:pk>/letteradd/<int:letter_pk>', views.manifestation_letter_add, name = 'manifestation_letter_add'),
+    path('manifestations/<int:pk>/letterremove/<int:letter_pk>', views.manifestation_letter_remove, name = 'manifestation_letter_remove'),
+    path('manifestations/<int:pk>/comment', views.ManifestationCommentUpdateView.as_view(), name = 'manifestation_comment'),
+    path('manifestations/<int:pk>/classification', views.ManifestationClassificationUpdateView.as_view(), name = 'manifestation_classification'),
+    path('manifestations/<int:pk>/print', views.manifestation_print_update, name = 'manifestation_print'),
+    path('manifestations/<int:pk>/print/add_publication_publisher/<int:publication_id>/<int:publisher_id>/', views.manifestation_add_publication_publisher, name = 'manifestation_add_publication_publisher'),
+    path('manifestations/<int:pk>/print/remove_publication_publisher/<int:publication_id>/', views.manifestation_remove_publication_publisher, name = 'manifestation_remove_publication_publisher'),
+    path('manifestations/<int:pk>/print/add_publication_place/<int:publication_id>/<int:place_id>/', views.manifestation_add_publication_place, name = 'manifestation_add_publication_place'),
+    path('manifestations/<int:pk>/print/remove_publication_place/<int:publication_id>/<int:place_id>', views.manifestation_remove_publication_place, name = 'manifestation_remove_publication_place'),
+    path('manifestations/publication/<int:pk>/delete/', views.ManifestationPublicationDeleteView.as_view(), name = 'manifestation_publication_delete'),
+    path('manifestations/<int:pk>/manuscript', views.manifestation_manuscript_update, name = 'manifestation_manuscript'),
+    path('manifestations/modification/<int:pk>/delete', views.ModificationDeleteView.as_view(), name='modification_delete'),
+    path('manifestations/modification/<int:modification_pk>/add_related_work/<int:work_pk>', views.modification_add_related_work, name='manifestation_modification_add_related_work'),
+    path('manifestations/modification/<int:modification_pk>/remove_related_work', views.modification_remove_related_work, name='manifestation_modification_remove_related_work'),
+    path('manifestations/modification/<int:modification_pk>/add_related_expression/<int:expression_pk>', views.modification_add_related_expression, name='manifestation_modification_add_related_expression'),
+    path('manifestations/modification/<int:modification_pk>/remove_related_expression', views.modification_remove_related_expression, name='manifestation_modification_remove_related_expression'),
+    path('manifestations/<int:pk>/digital_copy/', views.manifestation_digital_copy, name='manifestation_digital_copy'),
+    path('manifestations/<int:pk>/digital_copy/add/', views.manifestation_digital_copy_add, name='manifestation_digital_copy_add'),
+    path('manifestations/digital_copy/delete/<int:pk>/', views.ManifestationDigitalCopyDeleteView.as_view(), name='manifestation_digital_copy_delete'),
+    path('manifestations/<int:pk>/handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.manifestation_add_handwriting_writer, name = 'manifestation_add_handwriting_writer'),
+    path('manifestations/<int:pk>/handwriting/<int:handwriting_pk>/remove_writer', views.manifestation_remove_handwriting_writer, name = 'manifestation_remove_handwriting_writer'),
+    path('manifestations/<int:pk>/add_publisher/<int:publisher_id>', views.manifestation_add_publisher, name = 'manifestation_add_publisher'),
+    path('manifestations/<int:pk>/remove_publisher', views.manifestation_remove_publisher, name = 'manifestation_remove_publisher'),
+    path('manifestations/<int:pk>/add_stitcher/<int:stitcher_id>', views.manifestation_add_stitcher, name = 'manifestation_add_stitcher'),
+    path('manifestations/<int:pk>/remove_stitcher', views.manifestation_remove_stitcher, name = 'manifestation_remove_stitcher'),
+    path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/add_dedicatee/<int:person_id>/', views.manifestation_person_dedication_add_dedicatee, name='manifestation_person_dedication_add_dedicatee'),
+    path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/remove_dedicatee/<int:person_id>', views.manifestation_person_dedication_remove_dedicatee, name='manifestation_person_dedication_remove_dedicatee'),
+    path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/add_dedicatee/<int:corporation_id>/', views.manifestation_corporation_dedication_add_dedicatee, name='manifestation_corporation_dedication_add_dedicatee'),
+    path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/remove_dedicatee/<int:corporation_id>', views.manifestation_corporation_dedication_remove_dedicatee, name='manifestation_corporation_dedication_remove_dedicatee'),
+    path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/add_place/<int:place_id>/', views.manifestation_person_dedication_add_place, name='manifestation_person_dedication_add_place'),
+    path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/add_place/<int:place_id>/', views.manifestation_corporation_dedication_add_place, name='manifestation_corporation_dedication_add_place'),
+    path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/remove_place/', views.manifestation_person_dedication_remove_place, name='manifestation_person_dedication_remove_place'),
+    path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/remove_place/', views.manifestation_corporation_dedication_remove_place, name='manifestation_corporation_dedication_remove_place'),
+    path('manifestations/<int:pk>/delete', views.ManifestationDeleteView.as_view(), name = 'manifestation_delete'),
+    path('manifestations/<int:pk>/addplace/<int:place_id>', views.manifestation_add_place_view, name = 'manifestation_place_add'),
+    path('manifestations/<int:pk>/remove_place/<int:place_id>', views.manifestation_remove_place_view, name = 'manifestation_remove_place'),
+    path('items', views.ItemListView.as_view(), name = 'item_list'),
+    path('items/search', views.ItemSearchView.as_view(), name = 'item_search'),
+    path('items/<int:pk>/update', views.item_update, name = 'item_update'),
+    path('items/<int:pk>/detail', views.ItemDetailView.as_view(), name = 'item_detail'),
+    path('items/<int:pk>/delete', views.ItemDeleteView.as_view(), name = 'item_delete'),
+    path('items/<int:pk>/set_template', views.item_set_template, name = 'item_set_template'),
+    path('items/<int:pk>/relations', views.ItemRelationsUpdateView.as_view(), name = 'item_relations'),
+    path('items/<int:pk>/relations/addrelatedmanifestation/<int:target_item>', views.RelatedItemAddView.as_view(), name = 'related_item_add'),
+    path('items/removerelatedmanifestation/<int:pk>', views.RelatedItemRemoveView.as_view(), name = 'related_item_remove'),
+    path('items/<int:pk>/contributors', views.ItemContributorsUpdateView.as_view(), name = 'item_contributors'),
+    path('items/<int:pk>/contributors/addcontributor/<int:person>', views.ItemContributorAddView.as_view(), name = 'item_contributor_add'),
+    path('items/removecontributor/<int:pk>', views.ItemContributorRemoveView.as_view(), name = 'item_contributor_remove'),
+    path('items/<int:pk>/provenance', views.item_provenance, name = 'item_provenance'),
 
-        path('manifestation/person_dedication/add/<int:pk>/', views.person_dedication_add, name='person_dedication_add'),
-        path('manifestation/corporation_dedication/add/<int:pk>/', views.corporation_dedication_add, name='corporation_dedication_add'),
-        path('manifestation/person_dedication/delete/<int:pk>/', views.manifestation_person_dedication_delete, name='manifestation_person_dedication_delete'),
-        path('manifestation/corporation_dedication/delete/<int:pk>/', views.manifestation_corporation_dedication_delete, name='manifestation_corporation_dedication_delete'),
-        path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/add_dedicatee/<int:person_id>/', views.manifestation_person_dedication_add_dedicatee, name='manifestation_person_dedication_add_dedicatee'),
-        path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/remove_dedicatee/<int:person_id>', views.manifestation_person_dedication_remove_dedicatee, name='manifestation_person_dedication_remove_dedicatee'),
-        path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/add_dedicatee/<int:corporation_id>/', views.manifestation_corporation_dedication_add_dedicatee, name='manifestation_corporation_dedication_add_dedicatee'),
-        path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/remove_dedicatee/<int:corporation_id>', views.manifestation_corporation_dedication_remove_dedicatee, name='manifestation_corporation_dedication_remove_dedicatee'),
-        path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/add_place/<int:place_id>/', views.manifestation_person_dedication_add_place, name='manifestation_person_dedication_add_place'),
-        path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/add_place/<int:place_id>/', views.manifestation_corporation_dedication_add_place, name='manifestation_corporation_dedication_add_place'),
-        path('manifestation/<int:pk>/person_dedication/<int:dedication_id>/remove_place/', views.manifestation_person_dedication_remove_place, name='manifestation_person_dedication_remove_place'),
-        path('manifestation/<int:pk>/corporation_dedication/<int:dedication_id>/remove_place/', views.manifestation_corporation_dedication_remove_place, name='manifestation_corporation_dedication_remove_place'),
-        path('manifestations/<int:pk>/delete', views.ManifestationDeleteView.as_view(), name = 'manifestation_delete'),
-        path('manifestations/<int:pk>/addplace/<int:place_id>', views.manifestation_add_place_view, name = 'manifestation_place_add'),
-        path('manifestations/<int:pk>/remove_place/<int:place_id>', views.manifestation_remove_place_view, name = 'manifestation_remove_place'),
-        path('items', views.ItemListView.as_view(), name = 'item_list'),
-        path('items/search', views.ItemSearchView.as_view(), name = 'item_search'),
-        path('items/<int:pk>/update', views.item_update, name = 'item_update'),
-        path('items/<int:pk>/detail', views.ItemDetailView.as_view(), name = 'item_detail'),
-        path('items/<int:pk>/delete', views.ItemDeleteView.as_view(), name = 'item_delete'),
-        path('items/<int:pk>/set_template', views.item_set_template, name = 'item_set_template'),
-        path('items/<int:pk>/relations', views.ItemRelationsUpdateView.as_view(), name = 'item_relations'),
-        path('items/<int:pk>/relations/addrelatedmanifestation/<int:target_item>', views.RelatedItemAddView.as_view(), name = 'related_item_add'),
-        path('items/removerelatedmanifestation/<int:pk>', views.RelatedItemRemoveView.as_view(), name = 'related_item_remove'),
-        path('items/<int:pk>/contributors', views.ItemContributorsUpdateView.as_view(), name = 'item_contributors'),
-        path('items/<int:pk>/contributors/addcontributor/<int:person>', views.ItemContributorAddView.as_view(), name = 'item_contributor_add'),
-        path('items/removecontributor/<int:pk>', views.ItemContributorRemoveView.as_view(), name = 'item_contributor_remove'),
-        path('items/<int:pk>/provenance', views.item_provenance, name = 'item_provenance'),
+    path('item/<int:pk>/digital_copy/', views.item_digital_copy, name = 'item_digital_copy'),
+    path('item/<int:item_id>/digital_copy/add/', views.item_digital_copy_add, name = 'item_digital_copy_add'),
+    path('item/digital_copy/delete/<int:pk>/', views.ItemDigitalCopyDeleteView.as_view(), name = 'item_digital_copy_delete'),
 
-        path('item/<int:pk>/digital_copy/', views.item_digital_copy, name = 'item_digital_copy'),
-        path('item/<int:item_id>/digital_copy/add/', views.item_digital_copy_add, name = 'item_digital_copy_add'),
-        path('item/digital_copy/delete/<int:pk>/', views.ItemDigitalCopyDeleteView.as_view(), name = 'item_digital_copy_delete'),
+    path('item/<int:pk>/comment/', views.ItemCommentUpdateView.as_view(), name='item_comment'),
+    path('item/<int:pk>/dedication/', views.item_dedication, name='item_dedication'),
 
-        path('item/<int:pk>/comment/', views.ItemCommentUpdateView.as_view(), name='item_comment'),
-        path('item/<int:pk>/dedication/', views.item_dedication, name='item_dedication'),
+    path('item/dedication/person/add/<int:pk>/', views.item_person_dedication_add, name='item_person_dedication_add'),
+    path('item/dedication/corporation/add/<int:pk>/', views.item_corporation_dedication_add, name='item_corporation_dedication_add'),
+    path('item/dedication/person/delete/<int:pk>/', views.item_person_dedication_delete, name='item_person_dedication_delete'),
+    path('item/dedication/corporation/delete/<int:pk>/', views.item_corporation_dedication_delete, name='item_corporation_dedication_delete'),
+    path('item/dedication/person/add-dedicatee/<int:pk>/<int:dedication_id>/<int:person_id>/', views.item_person_dedication_add_dedicatee, name='item_person_dedication_add_dedicatee'),
+    path('item/dedication/person/remove-dedicatee/<int:pk>/<int:dedication_id>/', views.item_person_dedication_remove_dedicatee, name='item_person_dedication_remove_dedicatee'),
+    path('item/dedication/corporation/add-dedicatee/<int:pk>/<int:dedication_id>/<int:corporation_id>/', views.item_corporation_dedication_add_dedicatee, name='item_corporation_dedication_add_dedicatee'),
+    path('item/dedication/corporation/remove-dedicatee/<int:pk>/<int:dedication_id>/', views.item_corporation_dedication_remove_dedicatee, name='item_corporation_dedication_remove_dedicatee'),
+    path('item/dedication/add-place/<int:pk>/<int:dedication_id>/<int:place_id>/', views.item_dedication_add_place, name='item_dedication_add_place'),
+    path('item/dedication/remove-place/<int:pk>/<int:dedication_id>/', views.item_dedication_remove_place, name='item_dedication_remove_place'),
+    path('items/<int:pk>/manuscript', views.item_manuscript_update, name = 'item_manuscript'),
+    path('items/modification_handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.modification_add_handwriting_writer, name='modification_add_handwriting_writer'),
+    path('modification_handwriting/<int:handwriting_pk>/remove_writer', views.modification_remove_handwriting_writer, name='modification_remove_handwriting_writer'),
+    path('items/modification/<int:modification_pk>/add_related_work/<int:work_pk>', views.modification_add_related_work, name='modification_add_related_work'),
+    path('items/modification/<int:modification_pk>/remove_related_work', views.modification_remove_related_work, name='modification_remove_related_work'),
+    path('items/modification/<int:modification_pk>/add_related_expression/<int:expression_pk>', views.modification_add_related_expression, name='modification_add_related_expression'),
+    path('items/modification/<int:modification_pk>/remove_related_expression', views.modification_remove_related_expression, name='modification_remove_related_expression'),
+    path('libraries', views.LibraryListView.as_view(), name = 'library_list'),
+    path('libraries/search', views.LibrarySearchView.as_view(), name = 'library_search'),
+    path('libraries/create', views.LibraryCreateView.as_view(), name = 'library_create'),
+    path('libraries/<int:pk>/update', views.LibraryUpdateView.as_view(), name = 'library_update'),
+    path('libraries/<int:pk>/delete', views.LibraryDeleteView.as_view(), name = 'library_delete'),
+    path('letters', views.LetterListView.as_view(), name = 'letter_list'),
+    path('letters/search', views.LetterSearchView.as_view(), name = 'letter_search'),
+    path('letters/create', views.LetterCreateView.as_view(), name = 'letter_create'),
+    path('letters/<int:pk>/update', views.letter_update, name = 'letter_update'),
+    path('letters/<int:pk>/delete', views.LetterDeleteView.as_view(), name = 'letter_delete'),
 
-        path('item/dedication/person/add/<int:pk>/', views.item_person_dedication_add, name='item_person_dedication_add'),
-        path('item/dedication/corporation/add/<int:pk>/', views.item_corporation_dedication_add, name='item_corporation_dedication_add'),
-        path('item/dedication/person/delete/<int:pk>/', views.item_person_dedication_delete, name='item_person_dedication_delete'),
-        path('item/dedication/corporation/delete/<int:pk>/', views.item_corporation_dedication_delete, name='item_corporation_dedication_delete'),
-        path('item/dedication/person/add-dedicatee/<int:pk>/<int:dedication_id>/<int:person_id>/', views.item_person_dedication_add_dedicatee, name='item_person_dedication_add_dedicatee'),
-        path('item/dedication/person/remove-dedicatee/<int:pk>/<int:dedication_id>/', views.item_person_dedication_remove_dedicatee, name='item_person_dedication_remove_dedicatee'),
-        path('item/dedication/corporation/add-dedicatee/<int:pk>/<int:dedication_id>/<int:corporation_id>/', views.item_corporation_dedication_add_dedicatee, name='item_corporation_dedication_add_dedicatee'),
-        path('item/dedication/corporation/remove-dedicatee/<int:pk>/<int:dedication_id>/', views.item_corporation_dedication_remove_dedicatee, name='item_corporation_dedication_remove_dedicatee'),
-        path('item/dedication/add-place/<int:pk>/<int:dedication_id>/<int:place_id>/', views.item_dedication_add_place, name='item_dedication_add_place'),
-        path('item/dedication/remove-place/<int:pk>/<int:dedication_id>/', views.item_dedication_remove_place, name='item_dedication_remove_place'),
-        path('items/<int:pk>/manuscript', views.item_manuscript_update, name = 'item_manuscript'),
-        #path('items/modification_handwriting/<int:pk>/delete', views.ModificationHandwritingDeleteView.as_view(), name='modification_handwriting_delete'),
-        path('items/modification_handwriting/<int:handwriting_pk>/add_writer/<int:person_pk>', views.modification_add_handwriting_writer, name='modification_add_handwriting_writer'),
-        path('modification_handwriting/<int:handwriting_pk>/remove_writer', views.modification_remove_handwriting_writer, name='modification_remove_handwriting_writer'),
-        path('items/modification/<int:modification_pk>/add_related_work/<int:work_pk>', views.modification_add_related_work, name='modification_add_related_work'),
-        path('items/modification/<int:modification_pk>/remove_related_work', views.modification_remove_related_work, name='modification_remove_related_work'),
-        path('items/modification/<int:modification_pk>/add_related_expression/<int:expression_pk>', views.modification_add_related_expression, name='modification_add_related_expression'),
-        path('items/modification/<int:modification_pk>/remove_related_expression', views.modification_remove_related_expression, name='modification_remove_related_expression'),
-        path('libraries', views.LibraryListView.as_view(), name = 'library_list'),
-        path('libraries/search', views.LibrarySearchView.as_view(), name = 'library_search'),
-        path('libraries/create', views.LibraryCreateView.as_view(), name = 'library_create'),
-        path('libraries/<int:pk>/update', views.LibraryUpdateView.as_view(), name = 'library_update'),
-        path('libraries/<int:pk>/delete', views.LibraryDeleteView.as_view(), name = 'library_delete'),
-        path('letters', views.LetterListView.as_view(), name = 'letter_list'),
-        path('letters/search', views.LetterSearchView.as_view(), name = 'letter_search'),
-        path('letters/create', views.LetterCreateView.as_view(), name = 'letter_create'),
-        path('letters/<int:pk>/update', views.letter_update, name = 'letter_update'),
-        path('letters/<int:pk>/delete', views.LetterDeleteView.as_view(), name = 'letter_delete'),
-
-        path('letters/<int:pk>/add_sender_person/<int:person_id>', views.letter_add_sender_person, name='letter_add_sender_person'),
-        path('letters/<int:pk>/remove_sender_person', views.letter_remove_sender_person, name='letter_remove_sender_person'),
-        path('letters/<int:pk>/add_receiver_person/<int:person_id>', views.letter_add_receiver_person, name='letter_add_receiver_person'),
-        path('letters/<int:pk>/remove_receiver_person', views.letter_remove_receiver_person, name='letter_remove_receiver_person'),
-        path('letters/<int:pk>/add_sender_corporation/<int:corporation_id>', views.letter_add_sender_corporation, name='letter_add_sender_corporation'),
-        path('letters/<int:pk>/remove_sender_corporation', views.letter_remove_sender_corporation, name='letter_remove_sender_corporation'),
-        path('letters/<int:pk>/add_receiver_corporation/<int:corporation_id>', views.letter_add_receiver_corporation, name='letter_add_receiver_corporation'),
-        path('letters/<int:pk>/remove_receiver_corporation', views.letter_remove_receiver_corporation, name='letter_remove_receiver_corporation'),
-        path('letters/<int:pk>/add_sender_place/<int:place_id>', views.letter_add_sender_place, name='letter_add_sender_place'),
-        path('letters/<int:pk>/remove_sender_place', views.letter_remove_sender_place, name='letter_remove_sender_place'),
-        path('letters/<int:pk>/add_receiver_place/<int:place_id>', views.letter_add_receiver_place, name='letter_add_receiver_place'),
-        path('letters/<int:pk>/remove_receiver_place', views.letter_remove_receiver_place, name='letter_remove_receiver_place'),
-        path('letters/<int:pk>/add_edition/<str:zotitem_id>', views.letter_add_edition, name='letter_add_edition'),
-        path('letters/<int:pk>/remove_edition/<str:zotitem_id>', views.letter_remove_edition, name='letter_remove_edition'),
-    ]
-
-"""
-for entity in entities:
-    entity_class = ''.join(e.capitalize() for e in entity.split('_'))
-    if entities[entity] == '':
-        related_class_id = ''
-    else:
-        related_class_id = f"/<int:{entities[entity]}_id>"
-
-    urlpatterns.append(
-        path(
-                f"{entity}/create{related_class_id}",
-                getattr(views, f"{entity_class}CreateView").as_view(),
-                name = f"{entity}_create"
-            )
-        )
-
-    for action in ['update', 'delete']:
-        urlpatterns.append(
-            path(
-                    f"{entity}/{action}/<int:pk>",
-                    getattr(views, f"{entity_class}{action.capitalize()}View").as_view(),
-                    name = f"{entity}_{action}"
-                )
-            )
-urlpatterns = [
-        path('', views.IndexView.as_view(), name = 'index'),
-        path('works/<int:pk>', views.WorkDetailView.as_view(), name = 'work_detail'),
-        path('works/update/<int:pk>', views.WorkUpdateView.as_view(), name = 'work_update'),
-        path('works/new', views.WorkCreateView.as_view(), name = 'work_create'),
-        path('works/delete/<int:pk>', views.WorkDeleteView.as_view(), name = 'work_delete'),
-        path('worktitles/new/<int:work_id>', views.WorkTitleCreateView.as_view(), name = 'work_title_create'),
-        path('worktitles/update/<int:pk>', views.WorkTitleUpdateView.as_view(), name = 'work_title_update'),
-        path('worktitles/delete/<int:pk>', views.WorkTitleDeleteView.as_view(), name = 'work_title_delete'),
-        path('relatedwork/new/<int:work_id>', views.RelatedWorkCreateView.as_view(), name = 'related_work_create'),
-        path('relatedwork/update/<int:pk>', views.RelatedWorkUpdateView.as_view(), name = 'related_work_update'),
-        path('relatedwork/delete/<int:pk>', views.RelatedWorkDeleteView.as_view(), name = 'related_work_delete'),
-        path('contributor/new/<int:work_id>', views.ContributorCreateView.as_view(), name = 'contributor_create'),
-        path('contributor/update/<int:pk>', views.ContributorUpdateView.as_view(), name = 'contributor_update'),
-        path('contributor/delete/<int:pk>', views.ContributorDeleteView.as_view(), name = 'contributor_delete'),
-        path('expression_period/new/<int:expression_id>', views.ExpressionPeriodCreateView.as_view(), name = 'expression_period_create'),
-        path('expression_period/update/<int:pk>', views.ExpressionPeriodUpdateView.as_view(), name = 'expression_period_update'),
-        path('expression_period/delete/<int:pk>', views.ExpressionPeriodDeleteView.as_view(), name = 'expression_period_delete'),
-        path('expression_title/new/<int:expression_id>', views.ExpressionTitleCreateView.as_view(), name = 'expression_title_create'),
-        path('expression_title/update/<int:pk>', views.ExpressionTitleUpdateView.as_view(), name = 'expression_title_update'),
-        path('expression_title/delete/<int:pk>', views.ExpressionTitleDeleteView.as_view(), name = 'expression_title_delete'),
-        path('expression_contributor/new/<int:expression_id>', views.ExpressionContributorCreateView.as_view(), name = 'expression_contributor_create'),
-        path('expression_contributor/update/<int:pk>', views.ExpressionContributorUpdateView.as_view(), name = 'expression_contributor_update'),
-        path('expression_contributor/delete/<int:pk>', views.ExpressionContributorDeleteView.as_view(), name = 'expression_contributor_delete'),
-        path('expression/new/<int:work_id>', views.ExpressionCreateView.as_view(), name = 'expression_create'),
-        path('expression/update/<int:pk>', views.ExpressionUpdateView.as_view(), name = 'expression_update'),
-        path('expression/delete/<int:pk>', views.ExpressionDeleteView.as_view(), name = 'expression_delete'),
-    ]
-"""
+    path('letters/<int:pk>/add_sender_person/<int:person_id>', views.letter_add_sender_person, name='letter_add_sender_person'),
+    path('letters/<int:pk>/remove_sender_person', views.letter_remove_sender_person, name='letter_remove_sender_person'),
+    path('letters/<int:pk>/add_receiver_person/<int:person_id>', views.letter_add_receiver_person, name='letter_add_receiver_person'),
+    path('letters/<int:pk>/remove_receiver_person', views.letter_remove_receiver_person, name='letter_remove_receiver_person'),
+    path('letters/<int:pk>/add_sender_corporation/<int:corporation_id>', views.letter_add_sender_corporation, name='letter_add_sender_corporation'),
+    path('letters/<int:pk>/remove_sender_corporation', views.letter_remove_sender_corporation, name='letter_remove_sender_corporation'),
+    path('letters/<int:pk>/add_receiver_corporation/<int:corporation_id>', views.letter_add_receiver_corporation, name='letter_add_receiver_corporation'),
+    path('letters/<int:pk>/remove_receiver_corporation', views.letter_remove_receiver_corporation, name='letter_remove_receiver_corporation'),
+    path('letters/<int:pk>/add_sender_place/<int:place_id>', views.letter_add_sender_place, name='letter_add_sender_place'),
+    path('letters/<int:pk>/remove_sender_place', views.letter_remove_sender_place, name='letter_remove_sender_place'),
+    path('letters/<int:pk>/add_receiver_place/<int:place_id>', views.letter_add_receiver_place, name='letter_add_receiver_place'),
+    path('letters/<int:pk>/remove_receiver_place', views.letter_remove_receiver_place, name='letter_remove_receiver_place'),
+    path('letters/<int:pk>/add_edition/<str:zotitem_id>', views.letter_add_edition, name='letter_add_edition'),
+    path('letters/<int:pk>/remove_edition/<str:zotitem_id>', views.letter_remove_edition, name='letter_remove_edition'),
+]
