+++ apps/edwoca/views/item.py
@@ -52,21 +52,6 @@ class ItemSearchView(EdwocaSearchView):
         return super().get_queryset().filter(manifestation_is_singleton = False)
 
 
-class ItemSignatureDeleteView(DeleteView):
-    model = ItemSignature
-
-    def post(self, request, *args, **kwargs):
-        response = super().post(self, request, *args, **kwargs)
-        self.object.item.manifestation.save()
-        return response
-
-    def get_success_url(self):
-        if self.object.item.manifestation.is_singleton:
-            return reverse_lazy('edwoca:manifestation_update', kwargs={'pk': self.object.item.manifestation.id})
-        else:
-            return reverse_lazy('edwoca:item_update', kwargs={'pk': self.object.item.id})
-
-
 def item_update(request, pk):
     item = get_object_or_404(EdwocaItem, pk=pk)
     item_form = ItemForm(request.POST or None, instance=item)
