+++ apps/edwoca/views/manifestation/__init__.py
@@ -1,0 +1,4 @@
+from .common import *
+from .manuscript import *
+from .title import *
+from .update import *
