+++ apps/edwoca/views/manifestation/common.py
@@ -1,15 +1,14 @@
 import re
 from haystack.query import SQ
-from ..forms.manifestation import *
+from ...forms.manifestation import *
 from calendar import monthrange
-from ..forms.item import SignatureForm, ItemDigitizedCopyForm, PersonProvenanceStationForm, CorporationProvenanceStationForm, ItemProvenanceCommentForm, NewItemSignatureFormSet, ItemManuscriptForm, ItemHandwritingForm, PersonProvenanceStationBibForm, CorporationProvenanceStationBibForm, PersonProvenanceFormSet, PersonProvenanceBibFormSet, CorporationProvenanceFormSet, CorporationProvenanceBibFormSet, PersonProvenanceStationWebReference, CorporationProvenanceStationWebReference, PersonProvenanceStationWebReferenceForm, CorporationProvenanceStationWebReferenceForm
-from ..forms.modification import ItemModificationForm, ModificationHandwritingForm
-from ..forms.publication import PublicationForm
-from ..forms.dedication import ManifestationPersonDedicationForm, ManifestationCorporationDedicationForm
-from ..models import Manifestation as EdwocaManifestation, Letter, Expression, Work, ItemModification, ModificationHandwriting
-from .base import *
-from ..models import ManifestationTitle, ManifestationTitleHandwriting, ItemDigitalCopy
-from .base import *
+from ...forms.item import SignatureForm, ItemDigitizedCopyForm, PersonProvenanceStationForm, CorporationProvenanceStationForm, ItemProvenanceCommentForm, NewItemSignatureFormSet, ItemManuscriptForm, ItemHandwritingForm, PersonProvenanceStationBibForm, CorporationProvenanceStationBibForm, PersonProvenanceFormSet, PersonProvenanceBibFormSet, CorporationProvenanceFormSet, CorporationProvenanceBibFormSet, PersonProvenanceStationWebReference, CorporationProvenanceStationWebReference, PersonProvenanceStationWebReferenceForm, CorporationProvenanceStationWebReferenceForm
+from ...forms.modification import ItemModificationForm, ModificationHandwritingForm
+from ...forms.publication import PublicationForm
+from ...forms.dedication import ManifestationPersonDedicationForm, ManifestationCorporationDedicationForm
+from ...models import Manifestation as EdwocaManifestation, Letter, Expression, Work, ItemModification, ModificationHandwriting
+from ..base import *
+from ...models import ManifestationTitle, ManifestationTitleHandwriting, ItemDigitalCopy
 from bib.models import ZotItem
 from django.forms import inlineformset_factory
 from django.shortcuts import get_object_or_404, redirect, render
@@ -232,82 +231,6 @@ def manifestation_create(request, publisher_pk=None):
     return render(request, 'edwoca/create_manifestation.html', context)
 
 
-def manifestation_update(request, pk):
-    manifestation = Manifestation.objects.get(id=pk)
-    context = {
-        'object': manifestation,
-        'entity_type': 'manifestation'
-    }
-
-    if request.method == 'POST':
-        manifestation_form = ManifestationForm(request.POST, instance=manifestation)
-
-        signature_forms = []
-        if manifestation.is_singleton:
-            item = manifestation.get_single_item()
-            signature_forms_valid = True
-
-            for signature in manifestation.get_single_item().signatures.all():
-                signature_form = SignatureForm(
-                        request.POST,
-                        instance = signature,
-                        prefix = f"signature-{signature.id}"
-                    )
-                signature_forms.append(signature_form)
-
-        if manifestation_form.is_valid() and all(s.is_valid() for s in signature_forms):
-            manifestation_form.save()
-            for signature in manifestation.get_single_item().signatures.all():
-                signature.status = ItemSignature.Status.FORMER
-                signature.save()
-            for signature_form in signature_forms:
-                signature_form.save()
-        else:
-            context['manifestation_form'] = manifestation_form
-            context['signature_forms'] = signature_forms
-
-            return render(request, 'edwoca/manifestation_update.html', context)
-
-        if 'add_signature' in request.POST:
-            status = ItemSignature.Status.CURRENT
-            if item.signatures.count():
-                status = ItemSignature.Status.FORMER
-            signature = ItemSignature.objects.create(
-                    item = item,
-                    status = status
-                )
-            signature_forms += [ SignatureForm(instance = signature, prefix = f"signature-{signature.id}") ]
-
-        context['signature_forms'] = signature_forms
-        context['library_search_form'] = SearchForm()
-
-        return redirect('edwoca:manifestation_update', pk = pk)
-    else:
-        expression_search_form = FramedSearchForm(request.GET or None, prefix='expression', placeholder = _('search expressions'))
-        context['expression_search_form'] = expression_search_form
-        if 'expression_link' in request.GET:
-            expression_link = get_object_or_404(Expression, pk = request.GET['expression_link'])
-            context['expression_link'] = expression_link
-        if expression_search_form.is_valid() and expression_search_form.cleaned_data.get('q'):
-            context['expression_query'] = expression_search_form.cleaned_data.get('q')
-            context['found_expressions'] = expression_search_form.search().models(Expression)
-
-        manifestation_form = ManifestationForm(instance=manifestation)
-        signature_forms = []
-        for signature in manifestation.get_single_item().signatures.all():
-            signature_form = SignatureForm(
-                    instance = signature,
-                    prefix = f"signature-{signature.id}"
-                )
-            signature_forms.append(signature_form)
-
-        context['signature_forms'] = signature_forms
-        context['library_search_form'] = SearchForm()
-        context['manifestation_form'] = manifestation_form
-
-    return render(request, 'edwoca/manifestation_update.html', context)
-
-
 def manifestation_item_create(request, pk):
     manifestation = get_object_or_404(Manifestation, pk = pk)
     Item.objects.create(manifestation = manifestation)
@@ -347,228 +270,6 @@ def manifestation_unset_missing(request, pk):
     manifestation.save()
     return redirect('edwoca:manifestation_update', pk = pk)
 
-def manifestation_title_create(request, pk):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    ManifestationTitle.objects.create(manifestation=manifestation)
-    return redirect('edwoca:manifestation_title', pk = pk)
-
-def manifestation_title_update(request, pk):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    context = {
-        'object': manifestation,
-        'entity_type': 'manifestation'
-    }
-
-    if request.method == 'POST':
-        # collect any invalid forms and create a global render method in case one is invalid
-        form = ManifestationTitleDedicationForm(request.POST, instance=manifestation)
-        if form.is_valid():
-            form.save()
-        context['form'] = form
-
-        print_form = ManifestationPrintForm(request.POST, instance=manifestation)
-        if print_form.is_valid():
-            print_form.save()
-        # Handle existing title forms
-        title_forms = []
-        for title_obj in manifestation.titles.all():
-            prefix = f'title_{title_obj.id}'
-            title_form = ManifestationTitleForm(request.POST, instance=title_obj, prefix=prefix)
-            if title_form.is_valid():
-                title_form.save()
-
-            # Handle existing ManifestationTitleHandwriting forms for this title
-            for handwriting_obj in title_obj.handwritings.all():
-                handwriting_prefix = f'title_handwriting_{handwriting_obj.id}'
-                data = request.POST.copy()
-                if f'{handwriting_prefix}-medium' not in data:
-                    data[f'{handwriting_prefix}-medium'] = handwriting_obj.medium
-                handwriting_form = ManifestationTitleHandwritingForm(data, instance=handwriting_obj, prefix=handwriting_prefix)
-                if handwriting_form.is_valid():
-                    handwriting_form.save()
-            title_forms.append(title_form)
-        context['title_forms'] = title_forms
-
-        # Handle new title form
-        #new_title_form = ManifestationTitleForm(request.POST, prefix='new_title')
-        #if new_title_form.is_valid() and new_title_form.has_changed():
-            #new_title = new_title_form.save(commit=False)
-            #new_title.manifestation = manifestation
-            #new_title.save()
-
-        # Handle adding new ManifestationTitleHandwriting
-        if 'add_title_handwriting' in request.POST:
-            title_id_to_add_handwriting = request.POST.get('add_title_handwriting_to_title_id')
-            if title_id_to_add_handwriting:
-                title_obj = get_object_or_404(ManifestationTitle, pk=title_id_to_add_handwriting)
-                ManifestationTitleHandwriting.objects.create(manifestation_title=title_obj)
-
-        # Handle existing PersonDedication forms
-        person_dedication_forms = []
-        for person_dedication in manifestation.manifestation_person_dedications.all():
-            prefix = f'person_dedication_{person_dedication.id}'
-            form = ManifestationPersonDedicationForm(request.POST, instance=person_dedication, prefix=prefix)
-            if form.is_valid():
-                form.save()
-            if f'{prefix}-calculate-machine-readable-date' in request.POST:
-                period = person_dedication.period
-                period.parse_display()
-                period.save()
-                form = ManifestationPersonDedicationForm(instance=person_dedication, prefix=prefix)
-            if f'{prefix}-clear-machine-readable-date' in request.POST:
-                period = person_dedication.period
-                period.not_before = None
-                period.not_after = None
-                period.inferred = False
-                period.assumed = False
-                period.save()
-                form = ManifestationPersonDedicationForm(instance=person_dedication, prefix=prefix)
-            person_dedication_forms.append(form)
-        context['person_dedication_forms'] = person_dedication_forms
-
-        # Handle existing CorporationDedication forms
-        corporation_dedication_forms = []
-        for corporation_dedication in manifestation.manifestation_corporation_dedications.all():
-            prefix = f'corporation_dedication_{corporation_dedication.id}'
-            form = ManifestationCorporationDedicationForm(request.POST, instance=corporation_dedication, prefix=prefix)
-            if form.is_valid():
-                form.save()
-            if f'{prefix}-calculate-machine-readable-date' in request.POST:
-                period = corporation_dedication.period
-                period.parse_display()
-                period.save()
-                form = ManifestationCorporationDedicationForm(instance=corporation_dedication, prefix=prefix)
-            if f'{prefix}-clear-machine-readable-date' in request.POST:
-                period = corporation_dedication.period
-                period.not_before = None
-                period.not_after = None
-                period.inferred = False
-                period.assumed = False
-                period.save()
-                form = ManifestationCorporationDedicationForm(instance=corporation_dedication, prefix=prefix)
-            corporation_dedication_forms.append(form)
-        context['corporation_dedication_forms'] = corporation_dedication_forms
-
-        return redirect('edwoca:manifestation_title', pk = manifestation.id)
-
-    else:
-        # Initialize forms for existing titles
-        form = ManifestationTitleDedicationForm(instance=manifestation)
-        context['form'] = form
-
-        title_forms = []
-        for title_obj in manifestation.titles.all():
-            prefix = f'title_{title_obj.id}'
-            title_form = ManifestationTitleForm(instance=title_obj, prefix=prefix) # Get the form instance
-
-            # Initialize forms for existing ManifestationTitleHandwriting for this title
-            handwriting_forms = []
-            for handwriting_obj in title_obj.handwritings.all():
-                handwriting_prefix = f'title_handwriting_{handwriting_obj.id}'
-                handwriting_forms.append(ManifestationTitleHandwritingForm(instance=handwriting_obj, prefix=handwriting_prefix))
-            title_form.handwriting_forms = handwriting_forms # Attach to the form instance
-
-            title_forms.append(title_form) # Append the form instance to the list
-
-        context['title_forms'] = title_forms
-
-        # Initialize forms for existing PersonDedication
-        person_dedication_forms = []
-        for person_dedication in manifestation.manifestation_person_dedications.all():
-            prefix = f'person_dedication_{person_dedication.id}'
-            person_dedication_forms.append(ManifestationPersonDedicationForm(instance=person_dedication, prefix=prefix))
-        context['person_dedication_forms'] = person_dedication_forms
-
-        # Initialize forms for existing CorporationDedication
-        corporation_dedication_forms = []
-        for corporation_dedication in manifestation.manifestation_corporation_dedications.all():
-            prefix = f'corporation_dedication_{corporation_dedication.id}'
-            corporation_dedication_forms.append(ManifestationCorporationDedicationForm(instance=corporation_dedication, prefix=prefix))
-        context['corporation_dedication_forms'] = corporation_dedication_forms
-
-    context['print_form'] = ManifestationPrintForm(instance=manifestation)
-
-    q_dedicatee = request.GET.get('dedicatee-q')
-    q_place = request.GET.get('place-q')
-
-    if q_dedicatee:
-        dedicatee_search_form = FramedSearchForm(request.GET, prefix='dedicatee', placeholder=_('search persons'))
-        if dedicatee_search_form.is_valid():
-            context['query_dedicatee'] = dedicatee_search_form.cleaned_data.get('q')
-            context['found_persons'] = dedicatee_search_form.search().models(Person)
-            context['found_corporations'] = dedicatee_search_form.search().models(Corporation)
-    else:
-        dedicatee_search_form = FramedSearchForm(prefix='dedicatee', placeholder=_('search persons'))
-
-    if q_place:
-        place_search_form = FramedSearchForm(request.GET, prefix='place', placeholder=_('search place'))
-        if place_search_form.is_valid():
-            context['query_place'] = place_search_form.cleaned_data.get('q')
-            context['found_places'] = place_search_form.search().models(Place)
-    else:
-        place_search_form = FramedSearchForm(prefix='place', placeholder=_('search place'))
-
-    context['dedicatee_search_form'] = dedicatee_search_form
-    context['place_search_form'] = place_search_form
-
-    if request.GET.get('person_dedication_id'):
-        context['person_dedication_id'] = int(request.GET.get('person_dedication_id'))
-    if request.GET.get('corporation_dedication_id'):
-        context['corporation_dedication_id'] = int(request.GET.get('corporation_dedication_id'))
-
-    search_form = FramedSearchForm(request.GET or None, placeholder=_('search persons'))
-    context['search_form'] = search_form
-
-    if search_form.is_valid() and search_form.cleaned_data.get('q'):
-        context['query'] = search_form.cleaned_data.get('q')
-        context[f"found_persons"] = search_form.search().models(Person)
-
-    return render(request, 'edwoca/manifestation_title.html', context)
-
-
-def manifestation_writer_add(request, pk, title_id, writer_id):
-    title = get_object_or_404(ManifestationTitle, pk=title_id)
-    writer = get_object_or_404(Person, pk=writer_id)
-    title.writer = writer
-    title.save()
-    return redirect(reverse('edwoca:manifestation_title', kwargs={'pk': pk}) + f'#title-modal-{title_id}')
-
-
-def manifestation_writer_remove(request, pk, title_id):
-    title = get_object_or_404(ManifestationTitle, pk=title_id)
-    title.writer = None
-    title.save()
-    return redirect(reverse('edwoca:manifestation_title', kwargs={'pk': pk}) + f'#title-modal-{title_id}')
-
-
-def manifestation_title_add_handwriting_writer(request, pk, title_handwriting_pk, person_pk):
-    title_handwriting = get_object_or_404(ManifestationTitleHandwriting, pk=title_handwriting_pk)
-    person = get_object_or_404(Person, pk=person_pk)
-    title_handwriting.writer = person
-    title_handwriting.save()
-    return redirect(reverse('edwoca:manifestation_title', kwargs={'pk': pk}))
-
-
-def manifestation_title_remove_handwriting_writer(request, pk, title_handwriting_pk):
-    title_handwriting = get_object_or_404(ManifestationTitleHandwriting, pk=title_handwriting_pk)
-    title_handwriting.writer = None
-    title_handwriting.save()
-    return redirect(reverse('edwoca:manifestation_title', kwargs={'pk': pk}))
-
-
-def manifestation_add_dedicatee(request, pk, person_id):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    person = get_object_or_404(Person, pk=person_id)
-    manifestation.dedicatees.add(person)
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_remove_dedicatee(request, pk, dedicatee_id):
-    dedicatee = get_object_or_404(Person, pk=dedicatee_id)
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    manifestation.dedicatees.remove(dedicatee)
-    return redirect('edwoca:manifestation_title', pk=pk)
-
 
 class ManifestationDeleteView(EntityMixin, DeleteView):
     model = EdwocaManifestation
@@ -576,25 +277,11 @@ class ManifestationDeleteView(EntityMixin, DeleteView):
     template_name = 'edwoca/delete.html'
 
 
-class ManifestationTitleDeleteView(DeleteView):
-    model = ManifestationTitle
-
-    def get_success_url(self):
-        return reverse_lazy('edwoca:manifestation_title', kwargs = {'pk': self.object.manifestation.id})
-
-
 class ManifestationHandwritingDeleteView(DeleteView):
     model = ItemHandwriting
 
     def get_success_url(self):
         return reverse_lazy('edwoca:manifestation_manuscript', kwargs={'pk': self.object.item.manifestation.id})
-
-
-class ManifestationTitleHandwritingDeleteView(DeleteView):
-    model = ManifestationTitleHandwriting
-
-    def get_success_url(self):
-        return reverse_lazy('edwoca:manifestation_title', kwargs={'pk': self.object.manifestation_title.manifestation.id})
 
 
 class RelatedManifestationAddView(RelatedEntityAddView):
@@ -1487,283 +1174,6 @@ class CorporationProvenanceStationDeleteView(DeleteView):
         return reverse('edwoca:item_provenance', kwargs={'pk': item.id})
 
 
-def manifestation_manuscript_update(request, pk):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    item = manifestation.items.first()
-    context = {
-        'object': manifestation,
-        'entity_type': 'manifestation'
-    }
-    item_handwriting_ids = [
-            handwriting.id
-            for handwriting
-            in manifestation.get_single_item().handwritings.all()
-        ]
-    modification_handwriting_ids = [
-            handwriting.id
-            for modification
-            in manifestation.get_single_item().modifications.all()
-            for handwriting
-            in modification.handwritings.all()
-        ]
-
-    if request.method == 'POST':
-        form = ItemManuscriptForm(request.POST, instance=item)
-        if form.is_valid():
-            form.save()
-        else:
-            return render(request, 'edwoca/manifestation_manuscript.html', context)
-        context['form'] = form
-
-        handwriting_forms = []
-        for handwriting in item.handwritings.all():
-            prefix = f'handwriting_{handwriting.id}'
-            handwriting_form = ItemHandwritingForm(request.POST, instance=handwriting, prefix=prefix)
-            if handwriting_form.is_valid():
-                handwriting_form.save()
-            handwriting_forms.append(handwriting_form)
-        context['handwriting_forms'] = handwriting_forms
-
-        modifications = []
-
-        for modification in item.modifications.all():
-            for handwriting in modification.handwritings.all():
-                prefix = f'modification_handwriting_{handwriting.id}'
-                handwriting_form = ModificationHandwritingForm(request.POST, instance=handwriting, prefix=prefix)
-                if handwriting_form.is_valid():
-                    handwriting_form.save()
-
-            prefix = f'modification_{modification.id}'
-            modification_form = ItemModificationForm(request.POST, instance=modification, prefix=prefix)
-            if modification_form.is_valid():
-                modification_form.save()
-
-            if f'{prefix}-calculate-machine-readable-date' in request.POST:
-                period = modification.period
-                period.parse_display()
-                period.save()
-                modification_form = ItemModificationForm(instance=modification, prefix=prefix)
-            if f'{prefix}-clear-machine-readable-date' in request.POST:
-                period = modification.period
-                period.not_before = None
-                period.not_after = None
-                period.assumed = False
-                period.inferred = False
-                period.save()
-                modification_form = ItemModificationForm(instance=modification, prefix=prefix)
-
-
-            handwriting_forms = []
-            for handwriting in modification.handwritings.all():
-                prefix = f'modification_handwriting_{handwriting.id}'
-                handwriting_form = ModificationHandwritingForm(instance=handwriting, prefix = prefix)
-                if handwriting_form.is_valid():
-                    handwriting_form.save()
-                handwriting_forms.append(handwriting_form)
-
-            modifications.append({
-                'form': modification_form,
-                'handwriting_forms': handwriting_forms
-            })
-
-        if 'add_handwriting' in request.POST:
-            ItemHandwriting.objects.create(item=item)
-
-        if 'add_modification' in request.POST:
-            modification = ItemModification.objects.create(item=item)
-            prefix = f'modification_{modification.id}'
-
-            modifications.append({
-                'form': ItemModificationForm(instance = modification, prefix = prefix),
-                'handwriting_forms': []
-            })
-
-        context['modifications'] = modifications
-
-        if 'add_modification_handwriting' in request.POST:
-            modification_id = request.POST.get('add_modification_handwriting')
-            modification = get_object_or_404(ItemModification, pk=modification_id)
-            ModificationHandwriting.objects.create(modification=modification)
-
-        if 'delete_modification' in request.POST:
-            modification_id = request.POST.get('delete_modification')
-            modification = get_object_or_404(ItemModification, pk=modification_id)
-            modification.delete()
-
-        open_modifications = []
-        for key in request.POST:
-            if key.startswith('collapse_modification_'):
-                mod_id = int(key.split('_')[-1])
-                open_modifications.append(mod_id)
-        request.session['open_modifications'] = open_modifications
-
-        return redirect('edwoca:manifestation_manuscript', pk = manifestation.id)
-
-    else:
-        open_modifications = request.session.get('open_modifications', [])
-        if 'open_modifications' in request.session:
-            del request.session['open_modifications']
-        context['open_modifications'] = open_modifications
-
-        for handwriting_id in item_handwriting_ids + modification_handwriting_ids:
-            prefix = f'handwriting-{handwriting_id}'
-            if f'{prefix}-q' in request.GET:
-                search_form = SearchForm(request.GET or None, prefix = prefix)
-                if search_form.is_valid() and search_form.cleaned_data.get('q'):
-                    context['query'] = search_form.cleaned_data.get('q')
-                    context['search_form'] = search_form
-                    context['handwriting_found_persons'] = {
-                            'handwriting_id': handwriting_id,
-                            'results': search_form.search().models(Person)
-                        }
-
-        form = ItemManuscriptForm(instance=item)
-        handwriting_forms = []
-        for handwriting in item.handwritings.all():
-            prefix = f'handwriting_{handwriting.id}'
-            handwriting_forms.append(ItemHandwritingForm(instance=handwriting, prefix=prefix))
-        context['handwriting_forms'] = handwriting_forms
-
-        modifications = []
-        for modification in item.modifications.all():
-            #modification.ensure_period()
-            prefix = f'modification_{modification.id}'
-            modification_form = ItemModificationForm(instance=modification, prefix=prefix)
-
-            handwriting_forms = []
-            for handwriting in modification.handwritings.all():
-                prefix = f'modification_handwriting_{handwriting.id}'
-                handwriting_forms.append(ModificationHandwritingForm(instance=handwriting, prefix=prefix))
-
-            modifications.append({
-                'form': modification_form,
-                'handwriting_forms': handwriting_forms
-            })
-
-        context['modifications'] = modifications
-
-    context['form'] = form
-    #context['search_form'] = search_form
-
-    if request.GET.get('handwriting_id'):
-        context['handwriting_id'] = int(request.GET.get('handwriting_id'))
-
-    if request.GET.get('modification_id'):
-        context['add_handwriting_for_modification'] = True
-        context['modification_id'] = int(request.GET.get('modification_id'))
-        if search_form.is_valid() and search_form.cleaned_data.get('q'):
-            context['query'] = search_form.cleaned_data.get('q')
-            context[f"modification_found_persons"] = search_form.search().models(Person)
-
-    if request.GET.get('modification_handwriting_id'):
-        context['modification_handwriting_id'] = int(request.GET.get('modification_handwriting_id'))
-
-    # Search forms for modifications
-    for key in request.GET:
-        if key.startswith('modification-') and key.endswith('-q'):
-            parts = key.split('-')
-            modification_id = int(parts[1])
-            model_name = parts[2]
-
-            context['modification_id'] = modification_id
-            if modification_id not in open_modifications:
-                open_modifications.append(modification_id)
-
-            query = request.GET.get(key)
-
-            search_form = SearchForm({'q': query})
-
-            if model_name == 'work':
-                context['query_work'] = query
-                context['found_works'] = search_form.search().models(Work)
-            elif model_name == 'expression':
-                context['query_expression'] = query
-                context['found_expressions'] = search_form.search().models(Expression)
-            elif model_name == 'manifestation':
-                context['query_manifestation'] = query
-                context['found_manifestations'] = search_form.search().models(Manifestation)
-
-    work_search_form = SearchForm(prefix='work')
-    expression_search_form = SearchForm(prefix='expression')
-    manifestation_search_form = SearchForm(prefix='manifestation')
-
-    context['work_search_form'] = work_search_form
-    context['expression_search_form'] = expression_search_form
-    context['manifestation_search_form'] = manifestation_search_form
-
-    if request.GET.get('modification_id'):
-        context['modification_id'] = int(request.GET.get('modification_id'))
-
-    return render(request, 'edwoca/manifestation_manuscript.html', context)
-
-
-def manifestation_add_handwriting_writer(request, pk, handwriting_pk, person_pk):
-    handwriting = get_object_or_404(ItemHandwriting, pk=handwriting_pk)
-    person = get_object_or_404(Person, pk=person_pk)
-    handwriting.writer = person
-    handwriting.save()
-    return redirect('edwoca:manifestation_manuscript', pk=pk)
-
-
-def manifestation_remove_handwriting_writer(request, pk, handwriting_pk):
-    handwriting = get_object_or_404(ItemHandwriting, pk=handwriting_pk)
-    handwriting.writer = None
-    handwriting.save()
-    return redirect('edwoca:manifestation_manuscript', pk=pk)
-
-
-class ModificationHandwritingDeleteView(DeleteView):
-    model = ModificationHandwriting
-
-    def get_success_url(self):
-        if self.object.modification.item.manifestation.is_singleton:
-            return reverse_lazy('edwoca:manifestation_manuscript', kwargs={'pk': self.object.modification.item.manifestation.id})
-        else:
-            return reverse_lazy('edwoca:item_manuscript', kwargs={'pk': self.object.modification.item.item.id})
-
-
-def modification_add_related_work(request, modification_pk, work_pk):
-    modification = get_object_or_404(ItemModification, pk=modification_pk)
-    work = get_object_or_404(Work, pk=work_pk)
-    modification.related_work = work
-    modification.save()
-    if modification.item.manifestation.is_singleton:
-        return redirect('edwoca:manifestation_manuscript', pk=modification.item.manifestation.id)
-    else:
-        return redirect('edwoca:item_manuscript', pk=modification.item.id)
-
-
-def modification_remove_related_work(request, modification_pk):
-    modification = get_object_or_404(ItemModification, pk=modification_pk)
-    modification.related_work = None
-    modification.save()
-    if modification.item.manifestation.is_singleton:
-        return redirect('edwoca:manifestation_manuscript', pk=modification.item.manifestation.id)
-    else:
-        return redirect('edwoca:item_manuscript', pk=modification.item.id)
-
-
-def modification_add_related_expression(request, modification_pk, expression_pk):
-    modification = get_object_or_404(ItemModification, pk=modification_pk)
-    expression = get_object_or_404(Expression, pk=expression_pk)
-    modification.related_expression = expression
-    modification.save()
-    if modification.item.manifestation.is_singleton:
-        return redirect('edwoca:manifestation_manuscript', pk=modification.item.manifestation.id)
-    else:
-        return redirect('edwoca:item_manuscript', pk=modification.item.id)
-
-
-def modification_remove_related_expression(request, modification_pk):
-    modification = get_object_or_404(ItemModification, pk=modification_pk)
-    modification.related_expression = None
-    modification.save()
-    if modification.item.manifestation.is_singleton:
-        return redirect('edwoca:manifestation_manuscript', pk=modification.item.manifestation.id)
-    else:
-        return redirect('edwoca:item_manuscript', pk=modification.item.id)
-
-
 def person_provenance_add_owner(request, pk, pps_id, person_id):
     pps = get_object_or_404(PersonProvenanceStation, pk=pps_id)
     person = get_object_or_404(Person, pk=person_id)
@@ -1940,97 +1350,6 @@ class ManifestationDigitalCopyDeleteView(DeleteView):
     def get_success_url(self):
         manifestation = self.object.item.manifestation
         return reverse('edwoca:manifestation_digital_copy',kwargs={'pk': manifestation.id})
-
-
-
-def person_dedication_add(request, pk):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    period = Period.objects.create()
-    ManifestationPersonDedication.objects.create(manifestation=manifestation, period=period)
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def corporation_dedication_add(request, pk):
-    manifestation = get_object_or_404(Manifestation, pk=pk)
-    period = Period.objects.create()
-    ManifestationCorporationDedication.objects.create(manifestation=manifestation, period=period)
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_person_dedication_delete(request, pk):
-    dedication = get_object_or_404(ManifestationPersonDedication, pk=pk)
-    manifestation_pk = dedication.manifestation.pk
-    dedication.delete()
-    return redirect('edwoca:manifestation_title', pk=manifestation_pk)
-
-
-def manifestation_corporation_dedication_delete(request, pk):
-    dedication = get_object_or_404(ManifestationCorporationDedication, pk=pk)
-    manifestation_pk = dedication.manifestation.pk
-    dedication.delete()
-    return redirect('edwoca:manifestation_title', pk=manifestation_pk)
-
-
-def manifestation_person_dedication_add_dedicatee(request, pk, dedication_id, person_id):
-    dedication = get_object_or_404(ManifestationPersonDedication, pk=dedication_id)
-    person = get_object_or_404(Person, pk=person_id)
-    dedication.dedicatee.add(person)
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_person_dedication_remove_dedicatee(request, pk, dedication_id, person_id):
-    dedication = get_object_or_404(ManifestationPersonDedication, pk=dedication_id)
-    person = get_object_or_404(Person, pk=person_id)
-    dedication.dedicatee.remove(person)
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_corporation_dedication_add_dedicatee(request, pk, dedication_id, corporation_id):
-    dedication = get_object_or_404(ManifestationCorporationDedication, pk=dedication_id)
-    corporation = get_object_or_404(Corporation, pk=corporation_id)
-    dedication.dedicatee.add(corporation)
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_corporation_dedication_remove_dedicatee(request, pk, dedication_id, corporation_id):
-    dedication = get_object_or_404(ManifestationCorporationDedication, pk=dedication_id)
-    corporation = get_object_or_404(Corporation, pk=corporation_id)
-    dedication.dedicatee.remove(corporation)
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_person_dedication_add_place(request, pk, dedication_id, place_id):
-    dedication = ManifestationPersonDedication.objects.get(pk=dedication_id)
-    place = get_object_or_404(Place, pk=place_id)
-    dedication.place = place
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_person_dedication_remove_place(request, pk, dedication_id):
-    dedication = ManifestationPersonDedication.objects.get(pk=dedication_id)
-    dedication.place = None
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_corporation_dedication_add_place(request, pk, dedication_id, place_id):
-    dedication = get_object_or_404(ManifestationCorporationDedication, pk=dedication_id)
-    place = get_object_or_404(Place, pk=place_id)
-    dedication.place = place
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
-
-
-def manifestation_corporation_dedication_remove_place(request, pk, dedication_id):
-    dedication = get_object_or_404(ManifestationCorporationDedication, pk=dedication_id)
-    dedication.place = None
-    dedication.save()
-    return redirect('edwoca:manifestation_title', pk=pk)
 
 
 def manifestation_copy(request, pk):
